- removed unneeded fileUrl - removed unneeded affidavitFileUrl - cleaned up download calls and methods - removed unneeded DocumentIF - moved fetchDocument() from legal services to document mixin - renamed fetchDocument -> downloadDocument and refactored it to work with future Legal API endpoint
