+++ src/interfaces/store-interfaces/state-interfaces/continuation-authorization-interface.ts
@@ -3,7 +3,6 @@ export interface ContinuationAuthorizationIF {
     file: File // only used by UI
     fileKey: string
     fileName: string
-    fileUrl: string // presigned URL for future deletion (ticket 21110)
   }>
   date: string // YYYY-MM-DD
   expiryDate?: string // YYYY-MM-DD
